Learnt about JProgress Bar
Learnt about JProgressBar, which is perfect as an hp bar in a video game. Anyways, JProgressBar is a bar that can increase or decrease, you can set the minimum and max amount it increases, and can use a while loop and one way of using it is using a try catch with thread.sleep to be able to move the progress bar with a counter variable.
     * Another way, is linking it to a task and once a task is completed it will increase. You can customize it like many other J features we have tried, for instance, can add text, change font of text, color foreground and background and many more.